+++ src/main/java/com/imchobo/sayren_back/SayrenBackApplication.java
@@ -3,9 +3,11 @@ package com.imchobo.sayren_back;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.scheduling.annotation.EnableAsync;
 
 @SpringBootApplication
 @EnableJpaAuditing
+@EnableAsync
 public class SayrenBackApplication {
   public static void main(String[] args) {
     SpringApplication.run(SayrenBackApplication.class, args);
