+++ src/main/java/com/imchobo/sayren_back/domain/member/service/MemberServiceImpl.java
@@ -1,5 +1,6 @@
 package com.imchobo.sayren_back.domain.member.service;
 
+import com.imchobo.sayren_back.domain.common.service.MailService;
 import com.imchobo.sayren_back.domain.common.util.MailUtil;
 import com.imchobo.sayren_back.domain.common.util.RedisUtil;
 import com.imchobo.sayren_back.domain.member.dto.MemberSignupDTO;
@@ -31,7 +32,7 @@ public class MemberServiceImpl implements MemberService {
   private final PasswordEncoder passwordEncoder;
   private final MemberProviderRepository memberProviderRepository;
   private final RedisUtil redisUtil;
-  private final MailUtil mailUtil;
+  private final MailService mailService;
 
   @Override
   @Transactional
@@ -53,7 +54,7 @@ public class MemberServiceImpl implements MemberService {
 
 
     memberRepository.save(entity);
-    mailUtil.emailVerification(entity.getEmail());
+    mailService.emailVerification(entity.getEmail());
   }
 
   @Override
