+++ src/test/java/com/imchobo/sayren_back/domain/common/util/MailUtilTest.java
@@ -1,5 +1,6 @@
 package com.imchobo.sayren_back.domain.common.util;
 
+import com.imchobo.sayren_back.domain.common.service.MailService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -10,14 +11,10 @@ class MailUtilTest {
   @Autowired
   MailUtil mailUtil;
 
-
   @Test
   void sendMail() {
     mailUtil.sendMail("manlubo11@gmail.com", "메일 테스트", "내용입니다.");
   }
 
-  @Test
-  void emailVerification() {
-    mailUtil.emailVerification("manlubo11@gmail.com");
-  }
+
 }
